+++ resources/lang/en.json
@@ -184,5 +184,86 @@
   },
   "select_lang": {
     "title": "Select Language"
+  },
+  "flag_input": {
+    "title": "Flag Selector",
+    "real": "Real Flags",
+    "custom": "Custom Flags",
+    "select_layer": "Select a Layer",
+    "max_layer_reached_1": "You've reached the maximum number of layers.",
+    "max_layer_reached_2": "Please remove some to add new ones.",
+    "preview": "Preview",
+    "apply": "Apply Custom Flag",
+    "layer": "Layer",
+    "fixed": "fixed",
+    "remove": "Remove",
+    "color": "Color",
+    "layers": {
+      "frame": "Frame",
+      "full": "Full Background",
+      "center_circle": "Center Circle",
+      "center_star": "Center Star",
+      "center_flower": "Center Flower",
+      "center_hline": "Horizontal Stripe",
+      "center_vline": "Vertical Stripe",
+      "flower_tc": "Top Center Flower",
+      "flower_tl": "Top Left Flower",
+      "flower_tr": "Top Right Flower",
+      "diag_bl": "Bottom Left Diagonal",
+      "diag_br": "Bottom Right Diagonal",
+      "half_l": "Left Half",
+      "half_r": "Right Half",
+      "half_t": "Top Half",
+      "half_b": "Bottom Half",
+      "triangle_t": "Top Triangle",
+      "triangle_b": "Bottom Triangle",
+      "triangle_l": "Left Triangle",
+      "triangle_r": "Right Triangle",
+      "triangle_tl": "Top Left Triangle",
+      "triangle_tr": "Top Right Triangle",
+      "triangle_bl": "Bottom Left Triangle",
+      "triangle_br": "Bottom Right Triangle",
+      "tricolor_l": "Left Tricolor",
+      "tricolor_c": "Center Tricolor",
+      "tricolor_r": "Right Tricolor",
+      "tricolor_t": "Top Tricolor",
+      "tricolor_m": "Middle Tricolor",
+      "tricolor_b": "Bottom Tricolor",
+      "mini_tr_tl": "Mini Top Left Triangle",
+      "mini_tr_tr": "Mini Top Right Triangle",
+      "mini_tr_bl": "Mini Bottom Left Triangle",
+      "mini_tr_br": "Mini Bottom Right Triangle",
+      "nato_emblem": "NATO Emblem",
+      "eu_star": "EU Stars",
+      "laurel_wreath": "Laurel Wreath",
+      "octagram": "Octagram",
+      "octagram_2": "Diagonal Octagram",
+      "rocket": "Rocket",
+      "rocket_mini": "Mini Rocket",
+      "ofm_2025": "OFM 2025",
+      "translator": "Translator",
+      "beta_tester": "Beta Tester",
+      "beta_tester_circle": "Beta Circle",
+      "admin_contributors": "Contributors Emblem",
+      "admin_shield": "Admin Shield",
+      "admin_shield_r": "Right Admin Shield",
+      "admin_evan": "Evan's Crown",
+      "og": "OG Emblem",
+      "og_plus": "OG+ Emblem"
+    },
+    "reason": {
+      "admin_evan": "Only Evan can use this layer",
+      "admin": "Admin only",
+      "contributors": "Contributors only",
+      "beta": "Beta testers only",
+      "supporters": "Supporters only",
+      "og": "OG players only",
+      "og100": "OG100 players only",
+      "translator": "Translator only",
+      "well_known": "Well-known players only",
+      "known": "Known players only",
+      "seen": "Seen players only",
+      "login": "Login required"
+    }
   }
 }
