+++ resources/lang/ja.json
@@ -176,5 +176,86 @@
   },
   "select_lang": {
     "title": "言語を選択"
+  },
+  "flag_input": {
+    "title": "旗選択",
+    "real": "実在の旗",
+    "custom": "カスタム旗",
+    "select_layer": "レイヤーを選択",
+    "max_layer_reached_1": "レイヤー数の上限に達しています。",
+    "max_layer_reached_2": "新しく追加するには、いくつか削除してください。",
+    "preview": "プレビュー",
+    "apply": "カスタム旗を適用",
+    "layer": "レイヤー",
+    "fixed": "固定",
+    "remove": "削除",
+    "color": "色",
+    "layers": {
+      "frame": "フレーム",
+      "full": "背景",
+      "center_circle": "中央の円",
+      "center_star": "中央の星",
+      "center_flower": "中央の花",
+      "center_hline": "横ストライプ",
+      "center_vline": "縦ストライプ",
+      "flower_tc": "上中央の花",
+      "flower_tl": "左上の花",
+      "flower_tr": "右上の花",
+      "diag_bl": "左下の斜線",
+      "diag_br": "右下の斜線",
+      "half_l": "左半分",
+      "half_r": "右半分",
+      "half_t": "上半分",
+      "half_b": "下半分",
+      "triangle_t": "上三角形",
+      "triangle_b": "下三角形",
+      "triangle_l": "左三角形",
+      "triangle_r": "右三角形",
+      "triangle_tl": "左上三角形",
+      "triangle_tr": "右上三角形",
+      "triangle_bl": "左下三角形",
+      "triangle_br": "右下三角形",
+      "tricolor_l": "三色旗（左）",
+      "tricolor_c": "三色旗（中央）",
+      "tricolor_r": "三色旗（右）",
+      "tricolor_t": "三色旗（上）",
+      "tricolor_m": "三色旗（中央）",
+      "tricolor_b": "三色旗（下）",
+      "mini_tr_tl": "ミニ左上三角形",
+      "mini_tr_tr": "ミニ右上三角形",
+      "mini_tr_bl": "ミニ左下三角形",
+      "mini_tr_br": "ミニ右下三角形",
+      "nato_emblem": "NATOエンブレム",
+      "eu_star": "EUの星",
+      "laurel_wreath": "月桂冠",
+      "octagram": "八芒星",
+      "octagram_2": "斜めの八芒星",
+      "rocket": "ロケット",
+      "rocket_mini": "ミニロケット",
+      "ofm_2025": "OFM 2025",
+      "translator": "翻訳者マーク",
+      "beta_tester": "ベータテスター",
+      "beta_tester_circle": "ベータサークル",
+      "admin_contributors": "貢献者エンブレム",
+      "admin_shield": "管理者の盾",
+      "admin_shield_r": "管理者の盾（右）",
+      "admin_evan": "Evan様の冠",
+      "og": "OGエンブレム",
+      "og_plus": "OG+エンブレム"
+    },
+    "reason": {
+      "admin_evan": "このレイヤーはEvanのみ使用できます",
+      "admin": "管理者専用",
+      "contributors": "貢献者専用",
+      "beta": "ベータテスター専用",
+      "supporters": "支援者専用",
+      "og": "OGプレイヤー専用",
+      "og100": "OG100プレイヤー専用",
+      "translator": "翻訳者専用",
+      "well_known": "有名プレイヤー専用",
+      "known": "知られているプレイヤー専用",
+      "seen": "確認済みプレイヤー専用",
+      "login": "ログインが必要です"
+    }
   }
 }
