+++ src/client/index.html
@@ -349,6 +349,7 @@
     <dark-mode-button></dark-mode-button>
     <user-setting></user-setting>
     <multi-tab-modal></multi-tab-modal>
+    <flag-input-modal></flag-input-modal>
     <div
       id="language-modal"
       class="fixed inset-0 bg-black bg-opacity-50 z-50 hidden flex justify-center items-center"
